pyproject.toml
[tool.poetry]
name = "optisense-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
uvicorn = "^0.32.0"
uvicorn-worker = "^0.2.0"
gunicorn = "^23.0.0"
envparse = "^0.2.0"
psycopg2 = "^2.9.10"
drf-spectacular = "^0.27.2"
faker = "^30.8.2"
django-cors-headers = "^4.6.0"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
